refactor: move command-specific code to new `commands` module
